Split server workload tabs and reuse deployable identities
